Refactoring: Tags management
Now is ajaxification done properly (at least I think ;) ). When is Tag
being removed, only 2 queries are performed (Query for Tag with
particular ID and Delete query).

Updating Tag is done in the same way. Query for Tag and Update query.

And only result of this particular Tag is redrawn and sent to the
browser.